Introduced Test-Utils for Presets
